+++ lib/src/extensions/flutter_extensions/theme_extension.dart
@@ -904,6 +904,42 @@ extension FHUThemeDataExtension on ThemeData {
         package: package,
         overflow: overflow,
       );
+
+  /// Modifies the current [ThemeData] to optionally remove various visual feedback effects.
+  ///
+  /// This method returns a copy of the original [ThemeData] with modifications based
+  /// on the provided boolean arguments:
+  ///
+  /// - `noSplashColor` (default: `true`): If true, sets `splashColor` to transparent.
+  /// - `noHighlightColor` (default: `true`): If true, sets `highlightColor` to transparent.
+  /// - `noHoverColor` (default: `true`): If true, sets `hoverColor` to transparent.
+  /// - `noDividerColor` (default: `true`): If true, sets `dividerColor` to transparent.
+  /// - `noFocusColor` (default: `true`): If true, sets `focusColor` to transparent.
+  /// - `noSplashFactory` (default: `true`): If true, replaces `splashFactory` with a custom implementation
+  ///   that prevents splash effects.
+  ///
+  /// You can selectively disable individual effects by setting the corresponding argument to `false`.
+  /// For example, to keep the splash color while removing other effects, you would call:
+  ///
+  /// ```dart
+  /// ThemeData newTheme = theme.withoutEffects(noSplashColor: false);
+  /// ```
+  ThemeData withoutEffects({
+    bool noSplashColor = true,
+    bool noHighlightColor = true,
+    bool noHoverColor = true,
+    bool noDividerColor = true,
+    bool noFocusColor = true,
+    bool noSplashFactory = true,
+  }) =>
+      copyWith(
+        splashColor: noSplashColor ? Colors.transparent : null,
+        highlightColor: noHighlightColor ? Colors.transparent : null,
+        hoverColor: noHoverColor ? Colors.transparent : null,
+        dividerColor: noDividerColor ? Colors.transparent : null,
+        focusColor: noFocusColor ? Colors.transparent : null,
+        splashFactory: noSplashFactory ? NoSplash.splashFactory : null,
+      );
 }
 
 extension FHUThemeModeEx on ThemeMode? {
@@ -931,4 +967,82 @@ extension FHUBrightnessExtension on Brightness? {
   bool get isDark => this == Brightness.dark;
 
   bool get isLight => this == Brightness.light;
+}
+
+/// A widget that optionally removes various visual feedback effects from its child subtree.
+///
+/// This widget wraps its child in a [Theme] with modified theme data. The modifications
+/// include optionally setting the following properties to transparent:
+///
+/// - `splashColor`
+/// - `highlightColor`
+/// - `hoverColor`
+/// - `dividerColor`
+/// - `focusColor`
+///
+/// Additionally, it optionally replaces the `splashFactory` with a custom implementation that
+/// prevents splash effects from being generated.
+///
+/// Use this widget to selectively disable visual feedback in specific parts of your
+/// application where such effects are not desired.
+class ThemeWithoutEffects extends StatelessWidget {
+  /// Creates a [ThemeWithoutEffects] widget.
+  ///
+  /// The [child] is the widget subtree from which visual feedback effects
+  /// should be removed.
+  ///
+  /// The following arguments are booleans that determine if a specific effect should be removed:
+  ///
+  /// - `noSplashColor` (default: `true`)
+  /// - `noHighlightColor` (default: `true`)
+  /// - `noHoverColor` (default: `true`)
+  /// - `noDividerColor` (default: `true`)
+  /// - `noFocusColor` (default: `true`)
+  /// - `noSplashFactory` (default: `true`)
+  const ThemeWithoutEffects({
+    required this.child,
+    this.noSplashColor = true,
+    this.noHighlightColor = true,
+    this.noHoverColor = true,
+    this.noDividerColor = true,
+    this.noFocusColor = true,
+    this.noSplashFactory = true,
+    super.key,
+  });
+
+  /// The widget subtree to apply the effect to.
+  final Widget child;
+
+  /// Whether to remove the `splashColor` effect.
+  final bool noSplashColor;
+
+  /// Whether to remove the `highlightColor` effect.
+  final bool noHighlightColor;
+
+  /// Whether to remove the `hoverColor` effect.
+  final bool noHoverColor;
+
+  /// Whether to remove the `dividerColor` effect.
+  final bool noDividerColor;
+
+  /// Whether to remove the `focusColor` effect.
+  final bool noFocusColor;
+
+  /// Whether to remove the `splashFactory` effect.
+  final bool noSplashFactory;
+
+  @override
+  Widget build(BuildContext context) {
+    return Theme(
+      data: context.themeData.withoutEffects(
+        noSplashColor: noSplashColor,
+        noHighlightColor: noHighlightColor,
+        noHoverColor: noHoverColor,
+        noDividerColor: noDividerColor,
+        noFocusColor: noFocusColor,
+        noSplashFactory: noSplashFactory,
+      ),
+      child: child,
+    );
+  }
 }