+++ lib/src/extensions/string_extensions/general_string_extensions.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 
 extension FHUNullSafeStringExtensions on String? {
+  // all string extensions could be found in the https://
   // String get withoutWhiteSpaces => isEmptyOrNull ? '' : this!.replaceAll(RegExp(r'\s+\b|\b\s'), '');
   Size get textSize {
     final textPainter = TextPainter(