+++ lib/src/value_notifier/widgets/widgets.dart
@@ -1,1 +1,2 @@
+export 'box.dart';
 export 'listenables_builder.dart';